refactor: convert wishlist page to a server component
